modularize plotting function
- also slightly adapt some of the parameters

+++ cellular_raza-examples/pool_model_pyo3/cr_pool_model.py
@@ -103,10 +103,11 @@ def get_elements_at_all_iterations(output_path: Path, element_path="cell_storage
     pool = mp.Pool(threads)
     all_iterations = get_all_iterations(output_path, element_path)
     result = list(tqdm.tqdm(pool.imap(__iter_to_elements, map(lambda iteration: (output_path, iteration, element_path), all_iterations)), total=len(all_iterations)))
+
     return pd.concat(result)
 
 
-def save_snapshot(output_path, iteration, overwrite=False):
+def _determine_image_save_path(output_path, iteration):
     # Save images in dedicated images folder
     save_folder = Path(output_path) / "images"
     # Create folder if it does not exist
@@ -115,74 +116,19 @@ def save_snapshot(output_path, iteration, overwrite=False):
     # Create save path from new save folder
     save_path = save_folder / "snapshot_{:08}.png".format(iteration)
 
-    # If the image is present we do not proceed unless the overwrite flag is active
-    if overwrite==False and os.path.isfile(save_path):
-        return None
+    return save_path
 
-    # Get simulation settings and particles at the specified iteration
-    simulation_settings = get_simulation_settings(output_path)
-    df_cells = get_elements_at_iter(output_path, iteration, element_path="cell_storage")
-    df_voxels = get_elements_at_iter(output_path, iteration, element_path="voxel_storage")
 
-    # Get positions as large numpy array
-    positions = np.array([np.array(x) for x in df_cells["element.cell.mechanics.pos"]])
-    s = np.array([x for x in df_cells["element.cell.interaction.cell_radius"]])
-    # c = np.array([x for x in df_cells["element.cell.cellular_reactions.intracellular_concentrations"]])[:,1]
-    c = ["#24398c" if x else "#8c2424" for x in df_cells["element.cell.cellular_reactions.species"]=="S1"]
-
-    # Create a new norm for generating colors from matplotlib colorscales
-    # norm = matplotlib.colors.Normalize(
-    #     vmin=0,
-    #     vmax=c.max(),
-    #     clip=True,
-    # )
-    # mapper = matplotlib.cm.ScalarMappable(norm=norm, cmap=matplotlib.cm.summer)
-    # c = mapper.to_rgba(c.max() - c)
-
+def _create_base_canvas(domain):
     # Define limits for domain from simulation settings
-    xlims = np.array([0.0, simulation_settings.domain.size])
-    ylims = np.array([0.0, simulation_settings.domain.size])
+    xlims = np.array([0.0, domain.size])
+    ylims = np.array([0.0, domain.size])
 
     # Define the overall size of the figure and adapt to if the domain was not symmetric
     figsize_x = 16
     figsize_y = (ylims[1]-ylims[0])/(xlims[1]-xlims[0])*figsize_x
 
     fig, ax = plt.subplots(figsize=(figsize_x, figsize_y))
-
-    # Plot rectangles for background
-    # Create color mapper for background
-    nutrients_min = -0.2*simulation_settings.domain.initial_concentrations[0]
-    nutrients_max = simulation_settings.domain.initial_concentrations[0]
-    norm2 = matplotlib.colors.Normalize(
-        vmin=nutrients_min,
-        vmax=nutrients_max,
-        clip=True
-    )
-    mapper2 = matplotlib.cm.ScalarMappable(norm=norm2, cmap=matplotlib.cm.grey)
-
-    # Helper function to plot a single rectangle patch onto the canvas
-    def plot_rectangle(entry):
-        x_min = entry["element.voxel.min"]
-        x_max = entry["element.voxel.max"]
-        conc = entry["element.voxel.extracellular_concentrations"][0]
-
-        xy = [x_min[0], x_min[1]]
-        dx = x_max[0] - x_min[0]
-        dy = x_max[1] - x_min[1]
-        color = mapper2.to_rgba(conc) if not np.isnan(conc) else "red"
-        rectangle = matplotlib.patches.Rectangle(xy, width=dx, height=dy, color=color)
-        ax.add_patch(rectangle)
-
-    # Applies the previously defined plot_rectangle function to all voxels.
-    df_voxels.apply(plot_rectangle, axis=1)
-
-    # Plot circles for bacteria
-    for pos, si, ci in zip(positions, s, c):
-        if si!=None:
-            circle = plt.Circle(pos, radius=si, facecolor=ci, edgecolor=ci)
-            ax.add_patch(circle)
-        else:
-            print("Warning: Skip drawing bacteria with None radius!")
 
     # Sets correct boundaries for our domain
     ax.set_xlim(xlims)
@@ -192,6 +138,37 @@ def save_snapshot(output_path, iteration, overwrite=False):
     ax.get_xaxis().set_visible(False)
     ax.get_yaxis().set_visible(False)
 
+    return fig, ax
+
+
+def _create_nutrient_voxel_color_mapping(domain):
+    # Plot rectangles for background
+    # Create color mapper for background
+    nutrients_min = -0.2*domain.initial_concentrations[0]
+    nutrients_max = domain.initial_concentrations[0]
+    norm2 = matplotlib.colors.Normalize(
+        vmin=nutrients_min,
+        vmax=nutrients_max,
+        clip=True
+    )
+    return matplotlib.cm.ScalarMappable(norm=norm2, cmap=matplotlib.cm.grey)
+
+
+# Helper function to plot a single rectangle patch onto the canvas
+def _plot_rectangle(entry, mapper, ax):
+    x_min = entry["element.voxel.min"]
+    x_max = entry["element.voxel.max"]
+    conc = entry["element.voxel.extracellular_concentrations"][0]
+
+    xy = [x_min[0], x_min[1]]
+    dx = x_max[0] - x_min[0]
+    dy = x_max[1] - x_min[1]
+    color = mapper.to_rgba(conc) if not np.isnan(conc) else "red"
+    rectangle = matplotlib.patches.Rectangle(xy, width=dx, height=dy, color=color)
+    ax.add_patch(rectangle)
+
+
+def _plot_labels(fig, ax, n_bacteria_1, n_bacteria_2):
     # Calculate labels and dimensions for length scale
     x_locs = ax.get_xticks()
     y_locs = ax.get_yticks()
@@ -199,27 +176,78 @@ def save_snapshot(output_path, iteration, overwrite=False):
     x_low = np.min(x_locs[x_locs>0])/4
     y_low = np.min(y_locs[y_locs>0])/4
     dy = (y_locs[1]-y_locs[0])/8
-    dx = (x_locs[1] - x_locs[0])/2
+    dx = (x_locs[1] - x_locs[0])/1.5
 
     # Plot three rectangles as a length scale
     rectangle1 = matplotlib.patches.Rectangle([x_low,                y_low], width=    dx, height=dy, facecolor="grey", edgecolor="black")
     rectangle2 = matplotlib.patches.Rectangle([x_low +     dx,       y_low], width=    dx, height=dy, facecolor="white", edgecolor="black")
     rectangle3 = matplotlib.patches.Rectangle([x_low +   2*dx,       y_low], width=    dx, height=dy, facecolor="grey", edgecolor="black")
+
+    # Plot two rectangles to display species count
     rectangle4 = matplotlib.patches.Rectangle([x_low,           dy + y_low], width=1.5*dx, height=dy, facecolor="white", edgecolor="black")
     rectangle5 = matplotlib.patches.Rectangle([x_low + 1.5*dx,  dy + y_low], width=1.5*dx, height=dy, facecolor="white", edgecolor="black")
+
+    # Add all rectangles to ax
     ax.add_patch(rectangle1)
     ax.add_patch(rectangle2)
     ax.add_patch(rectangle3)
     ax.add_patch(rectangle4)
     ax.add_patch(rectangle5)
+
+    # Display size of the
     ax.annotate(f"{dx:5.0f}µm", (x_low +   dx/2, y_low+dy/2), ha='center', va='center')
     ax.annotate(f"{dx:5.0f}µm", (x_low + 3*dx/2, y_low+dy/2), ha='center', va='center')
     ax.annotate(f"{dx:5.0f}µm", (x_low + 5*dx/2, y_low+dy/2), ha='center', va='center')
+
     # Create rectangle to show number of agents
-    n_bacteria_1 = np.sum(df_cells["element.cell.cellular_reactions.species"]=="S1")
-    n_bacteria_2 = len(positions) - n_bacteria_1
     ax.annotate(f"Species 1 {n_bacteria_1:5.0f}", (x_low+1.5*dx/2, y_low+3*dy/2), ha='center', va='center')
     ax.annotate(f"Species 2 {n_bacteria_2:5.0f}", (x_low+4.5*dx/2, y_low+3*dy/2), ha='center', va='center')
+
+
+def _plot_voxels(df_voxels, ax, mapper):
+    # Applies the previously defined plot_rectangle function to all voxels.
+    df_voxels.apply(lambda entry: _plot_rectangle(entry, mapper, ax), axis=1)
+
+
+def _plot_bacteria(df_cells, ax):
+    # Get positions as large numpy array
+    positions = np.array([np.array(x) for x in df_cells["element.cell.mechanics.pos"]])
+    s = np.array([x for x in df_cells["element.cell.interaction.cell_radius"]])
+    c = ["#24398c" if x else "#8c2424" for x in df_cells["element.cell.cellular_reactions.species"]=="S1"]
+
+    # Plot circles for bacteria
+    for pos, si, ci in zip(positions, s, c):
+        if si!=None:
+            circle = plt.Circle(pos, radius=si, facecolor=ci, edgecolor=ci)
+            ax.add_patch(circle)
+        else:
+            print("Warning: Skip drawing bacteria with None radius!")
+
+
+def save_snapshot(output_path, iteration, overwrite=False):
+    save_path = _determine_image_save_path(output_path, iteration)
+
+    # If the image is present we do not proceed unless the overwrite flag is active
+    if overwrite==False and os.path.isfile(save_path):
+        return None
+
+    # Get simulation settings and particles at the specified iteration
+    domain, cells, meta_params = get_simulation_settings(output_path)
+    df_cells = get_elements_at_iter(output_path, iteration, element_path="cell_storage")
+    df_voxels = get_elements_at_iter(output_path, iteration, element_path="voxel_storage")
+
+    fig, ax = _create_base_canvas(domain)
+
+    mapper2 = _create_nutrient_voxel_color_mapping(domain)
+
+    _plot_voxels(df_voxels, ax, mapper2)
+
+    _plot_bacteria(df_cells, ax)
+
+    # Plot labels in bottom left corner
+    n_bacteria_1 = len(df_cells[df_cells["element.cell.cellular_reactions.species"]=="S1"])
+    n_bacteria_2 = len(df_cells[df_cells["element.cell.cellular_reactions.species"]!="S1"])
+    _plot_labels(fig, ax, n_bacteria_1, n_bacteria_2)
 
     # Save figure and cut off excess white space
     fig.savefig(save_path, bbox_inches='tight', pad_inches = 0)
@@ -228,7 +256,6 @@ def save_snapshot(output_path, iteration, overwrite=False):
     plt.close(fig)
     del df_cells
     del df_voxels
-    del positions
     del fig
     del ax
     gc.collect()
